requirements.txt
asgiref==3.11.1
attrs==26.1.0
defusedxml==0.7.1
diff-match-patch==20241021
Django==6.0.3
django-import-export==4.4.0
django-json-rpc==0.7.2
django-modern-rpc==2.0.0
djangorestframework==3.17.1
json-rpc==1.15.0
jsonrpcserver==5.0.9
jsonrpyc==1.3.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
OSlash==0.6.3
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
tablib==3.9.0
typing_extensions==4.15.0
tzdata==2026.1
